Add the effects of CA_chap_seign
The effects of CA_chap_seign cannot be given by a building.
I have uncommented them to indicate to the player what will be the
effects and the effects are given by a character modifier.
The management of this modifier is done by on_action events for the
construction and the holder change.