Manage better the new holder of FEODAL_adm buildings
When there is a new holder, the province whose the capital has some
FEODAL_adm buildings can stay the capital or not.
In the events triggered by the on_actions, we cannot check if the title
is the capital or not (maybe because the capital is computed after).
So I propose to give the modifiers and to remove them quickly if in fact
we shouldn't give them. For this I have triggered the removal event.