Hide set_adm_spe when we have already the adm spe
I have moved some triggers from the potential to the allow bloc in order
to know why we cannot set an admin specialization (7d89def)

But when we have already the admin spe, the decision appears for
nothing.
So I manage this case by checking in potential that the specialization
of the capital is mil ou prod to display the decision (also manage when
there is no specialization).