Add event to spe faster new holding
When an holding is built, we give it directly the event to be
specialized.

Previously, there was a short MTTH for this case. But it caused some
bugs with tribal to feudal (or republic) conversion.
That's why Arko change the MTTH in 34d85bcf and we need an on_action
event for the new holdings